requirements.txt
apiai
asn1crypto
certifi
cffi
cryptography
DateTime
future
numpy
Pillow
pycparser
python-telegram-bot
pytz
six
telegram
zope.interface
django
gunicorn
django-heroku